validation completed

+++ src/main/java/com/lawrence/expensetrackerapi/service/Impl/ExpenseServiceImpl.java
@@ -1,6 +1,7 @@
 package com.lawrence.expensetrackerapi.service.Impl;
 
 import com.lawrence.expensetrackerapi.entity.Expense;
+import com.lawrence.expensetrackerapi.exceptions.ResourceNotFoundException;
 import com.lawrence.expensetrackerapi.repository.ExpenseRepository;
 import com.lawrence.expensetrackerapi.service.ExpenseService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,12 +29,16 @@ public class ExpenseServiceImpl implements ExpenseService {
         if(expense.isPresent()){
             return expense.get();
         }
-        throw  new RuntimeException("Expense is not found for the id:" + id);
+        throw new ResourceNotFoundException("Expense with the following id:" + id + " not found!");
     }
 
     @Override
     public void deleteExpenseById(Long id) {
-           expenseRepo.deleteById(id);
+        Optional<Expense> expense = expenseRepo.findById(id);
+        if(expense.isPresent()){
+            expenseRepo.deleteById(id);
+        }
+        throw new ResourceNotFoundException("Expense with the following id:" + id + " not found!");
     }
 
     @Override
@@ -49,5 +55,27 @@ public class ExpenseServiceImpl implements ExpenseService {
         existingExpense.setDate(expense.getDate() != null ? expense.getDate() : existingExpense.getDate());
         existingExpense.setAmount(expense.getAmount() != null ? expense.getAmount() : existingExpense.getAmount());
         return expenseRepo.save(existingExpense);
+    }
+
+    @Override
+    public List<Expense> readByCategory(String category, Pageable page) {
+        return expenseRepo.findByCategory(category, page).toList();
+    }
+
+    @Override
+    public List<Expense> readByName(String name, Pageable page) {
+        return expenseRepo.findByNameContaining(name, page).toList();
+    }
+
+    @Override
+    public List<Expense> readByDate(Date startDate, Date endDate, Pageable page) {
+        if(startDate == null){
+            startDate = new Date(0);
+        }
+        if(endDate == null){
+            endDate = new Date(System.currentTimeMillis());
+        }
+        Page<Expense> pages = expenseRepo.findByDateBetween(startDate, endDate, page);
+        return pages.toList();
     }
 }
